requirements.txt
Pillow
pytesseract
psutil